requirements.txt
#django
#djangorestframework
#django-cors-headers
#python-keycloak
#python-keycloak-client
#asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
deprecation==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
ecdsa==0.18.0
httpie==3.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
openapi-codec==1.3.2
packaging==23.1
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
PyJWT==1.7.1
PySocks==1.7.1
python-jose==3.3.0
python-keycloak==3.3.0
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.3
rsa==4.9
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4